+++ CHANGELOG.md
@@ -28,6 +28,7 @@
     - Add `Action.dryMatch`.
 - Parser
   - **_Breaking Change_**: add `ASTNode.token`.
+  - **_Breaking Change_**: `IParserBuilder.useLexer` must be called and only can be called once to set the lexer, instead of setting the lexer in `IParserBuilder.build`.
   - Feat: add `IParserBuilder.useData`.
 - Optimize package size.